zmart-foundation/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.29.0
httpx==0.27.2
redis==5.0.8
python-dotenv==1.0.1
loguru==0.7.2

opentelemetry-sdk==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-exporter-otlp==1.26.0

APScheduler==3.10.4
pytest==8.3.2
pandas==2.2.3
